+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * Added:
   * `bump` option.
+Fixed:
+  * `poetry shell` did not clean up after itself.
 
 ## v0.10.0 (2020-10-08)